+++ packages/create-yoshi-app/src/auto-release/setupAutoRelease.ts
@@ -12,12 +12,13 @@ export default async (templateModel: TemplateModel) => {
     await updateAppModel(templateModel.projectName, config);
     console.log(
       chalk.cyan(
-        `\nAuto-release for ${chalk.underline(
-          templateModel.projectName,
-        )} was configured ✅`,
+        `${chalk.hex('66CF9C')('✔')} ${chalk.bold(
+          'Auto-release',
+        )} for ${chalk.underline(templateModel.projectName)} was configured`,
       ),
     );
   } catch (e) {
+    console.error(e);
     console.error(
       `Can't set up an auto release for current project.
 Please check the docs to configure it manually: