+++ packages/create-yoshi-app/src/dev-center-registration/runPrompt.ts
@@ -3,24 +3,32 @@ import extendedPropmts, { Answers } from '../extended-prompts';
 import { initAPIService } from '../api';
 import TemplateModel from './TemplateModel';
 import getQuestions from './getQuestions';
-import { getInstance } from './auth';
+import { getAuthInstance } from './auth';
 import getDefaultAnswers from './getDefaultAnswers';
+
+const getFallback = (localAppModel: LocalAppTemplateModel) => () => {
+  console.log(
+    "⚠ Can't register/migrate Dev Center project. The default project will be created instead.",
+    '\n',
+  );
+
+  return new TemplateModel(
+    getDefaultAnswers(localAppModel.templateDefinition.name),
+  );
+};
 
 export default async (
   localAppModel: LocalAppTemplateModel,
+  questions = getQuestions(),
+  fallback = getFallback(localAppModel),
 ): Promise<TemplateModel> => {
-  const fallbackWithDefaultTemplate = () =>
-    new TemplateModel(getDefaultAnswers(localAppModel.templateDefinition.name));
-
-  const instance = await getInstance();
+  const instance = await getAuthInstance();
   if (instance) {
     initAPIService(instance);
   } else {
     // We should handle default template if auth flow was failed
-    return fallbackWithDefaultTemplate();
+    return fallback();
   }
-
-  const questions = getQuestions();
 
   let answers: Answers<string>;
 
@@ -30,17 +38,13 @@ export default async (
       localAppModel,
     );
     if (!answers.appRegistrationState) {
-      return fallbackWithDefaultTemplate();
+      return fallback();
     }
   } catch (e) {
     // We want to show unhandled errors
     if (e.message !== 'Aborted') {
       console.error(e);
-      console.log(
-        "⚠ Can't register/migrate Dev Center project. The default project will be created instead.",
-        '\n',
-      );
-      return fallbackWithDefaultTemplate();
+      return fallback();
     }
     console.log();
     console.log('Aborting ...');