+++ packages/yoshi-flow-bm-runtime/src/wrapComponent.tsx
@@ -1,16 +1,91 @@
-import React, { ComponentType } from 'react';
+import React, { ComponentType, useMemo } from 'react';
 import ModuleProvider, { IBMModuleParams } from './hooks/ModuleProvider';
 
-export default function wrapComponent(
-  Component: ComponentType,
-  deps: Array<ComponentType>,
-): ComponentType<IBMModuleParams> {
-  const children = deps.reduce(
-    (children, Provider) => <Provider>{children}</Provider>,
-    <Component />,
-  );
+interface AdditionalProps {
+  routeBaseName?: string;
+  router?: any;
+}
 
-  return props => (
-    <ModuleProvider moduleParams={props}>{children}</ModuleProvider>
-  );
+export default function wrapComponent<P extends {}>(
+  Component: ComponentType<P>,
+  deps: Array<ComponentType>,
+): ComponentType<IBMModuleParams & AdditionalProps & P> {
+  return props => {
+    const {
+      routeBaseName,
+      router,
+      metaSiteId,
+      children,
+      config,
+      accountLanguage,
+      brand,
+      coBranding,
+      debug,
+      instance,
+      instanceId,
+      liveSite,
+      locale,
+      primarySiteLocale,
+      siteName,
+      userId,
+      userPermissions,
+      userRole,
+      viewMode,
+      ...restProps
+    } = props;
+
+    const moduleParams: IBMModuleParams & AdditionalProps = useMemo(
+      () => ({
+        routeBaseName,
+        router,
+        metaSiteId,
+        config,
+        accountLanguage,
+        brand,
+        coBranding,
+        debug,
+        instance,
+        instanceId,
+        liveSite,
+        locale,
+        primarySiteLocale,
+        siteName,
+        userId,
+        userPermissions,
+        userRole,
+        viewMode,
+      }),
+      [
+        routeBaseName,
+        router,
+        metaSiteId,
+        config,
+        accountLanguage,
+        brand,
+        coBranding,
+        debug,
+        instance,
+        instanceId,
+        liveSite,
+        locale,
+        primarySiteLocale,
+        siteName,
+        userId,
+        userPermissions,
+        userRole,
+        viewMode,
+      ],
+    );
+
+    return (
+      <ModuleProvider moduleParams={moduleParams}>
+        {deps.reduce(
+          (children, Provider) => (
+            <Provider>{children}</Provider>
+          ),
+          <Component {...((restProps as unknown) as P)}>{children}</Component>,
+        )}
+      </ModuleProvider>
+    );
+  };
 }