+++ packages/yoshi-flow-bm/src/config/types.ts
@@ -3,6 +3,7 @@ export type InitialModuleConfig = Partial<ModuleConfig>;
 export interface ModuleConfig {
   moduleId: string;
   moduleConfigurationId?: string;
+  appDefId?: string;
   experimentsScopes: Array<string>;
   sentryDsn?: string;
   routeNamespace: string;