+++ packages/yoshi-flow-monorepo/src/webpack.config.ts
@@ -15,10 +15,13 @@ import {
   isSingleEntry,
   inTeamCity,
   isProduction as isProductionQuery,
+  inMasterTeamCity,
 } from 'yoshi-helpers/build/queries';
 import { STATICS_DIR, SERVER_ENTRY, SRC_DIR } from 'yoshi-config/build/paths';
 import ManifestPlugin from 'yoshi-common/build/manifest-webpack-plugin';
 import { isObject } from 'lodash';
+import SentryWebpackPlugin from '@sentry/webpack-plugin';
+import { getProjectArtifactVersion } from 'yoshi-helpers/utils';
 import { PackageGraphNode } from './load-package-graph';
 import { isThunderboltElementModule, isThunderboltAppModule } from './utils';
 
@@ -112,6 +115,17 @@ export function createClientWebpackConfig(
 
   if (isThunderboltElementModule(pkg)) {
     clientConfig.optimization!.runtimeChunk = false;
+  }
+
+  if (isThunderboltAppModule(pkg)) {
+    if (inMasterTeamCity()) {
+      clientConfig.plugins!.push(
+        new SentryWebpackPlugin({
+          include: path.join(pkg.location, STATICS_DIR),
+          release: getProjectArtifactVersion(),
+        }),
+      );
+    }
   }
 
   clientConfig.entry = isSingleEntry(entry) ? { app: entry as string } : entry;
@@ -273,6 +287,10 @@ export function createWebWorkerWebpackConfig(
   workerConfig.resolve!.alias = pkg.config.webWorkerResolveAlias;
   workerConfig.externals = pkg.config.webWorkerExternals;
 
+  if (pkg.config.webWorkerSplitChunks) {
+    workerConfig.optimization!.splitChunks = pkg.config.webWorkerSplitChunks;
+  }
+
   return workerConfig;
 }
 
@@ -293,6 +311,7 @@ export function createWebWorkerServerWebpackConfig(
     isHot,
     isMonorepo: true,
     createWorkerManifest: false,
+    overrideDefinePluginBrowserEnvVar: false,
     ...defaultOptions,
   });
 
@@ -323,6 +342,8 @@ export function createSiteAssetsWebpackConfig(
     isAnalyze,
     target,
     transpileCarmiOutput,
+    disableEmitSourceMaps = false,
+    forceMinimizeServer = false,
   }: {
     isDev?: boolean;
     forceEmitSourceMaps?: boolean;
@@ -330,6 +351,8 @@ export function createSiteAssetsWebpackConfig(
     isAnalyze?: boolean;
     target: 'web' | 'node';
     transpileCarmiOutput?: boolean;
+    disableEmitSourceMaps?: boolean;
+    forceMinimizeServer?: boolean;
   },
 ): webpack.Configuration {
   const entry = pkg.config.entry || defaultEntry;
@@ -348,6 +371,8 @@ export function createSiteAssetsWebpackConfig(
     // We don't have any server externals for `site assets` bundle
     // So with empty object, we'll be sure that no default externals value will be applied
     serverExternals: target === 'node' ? {} : undefined,
+    forceMinimizeServer,
+    disableEmitSourceMaps,
     exportAsLibraryName: pkg.config.exports,
     enhancedTpaStyle: pkg.config.enhancedTpaStyle,
     tpaStyle: pkg.config.tpaStyle,